+++ packages/builder/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @openzeppelin/transaction-form-builder-core
 
+## 0.1.5
+
+### Patch Changes
+
+- [#57](https://github.com/OpenZeppelin/contracts-ui-builder/pull/57) [`fc05836`](https://github.com/OpenZeppelin/contracts-ui-builder/commit/fc058368cbd885b63c5aa932ed216df783f59e71) Thanks [@pasevin](https://github.com/pasevin)! - resolve network switching regression and optimize state management
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [[`83c430e`](https://github.com/OpenZeppelin/contracts-ui-builder/commit/83c430e86f47733bde89b560b70a7a922eebfe81), [`83c430e`](https://github.com/OpenZeppelin/contracts-ui-builder/commit/83c430e86f47733bde89b560b70a7a922eebfe81)]:
+  - @openzeppelin/contracts-ui-builder-adapter-evm@0.2.0
+  - @openzeppelin/contracts-ui-builder-styles@0.2.0
+  - @openzeppelin/contracts-ui-builder-types@0.2.0
+  - @openzeppelin/contracts-ui-builder-utils@0.2.0
+  - @openzeppelin/contracts-ui-builder-ui@0.2.0
+  - @openzeppelin/contracts-ui-builder-adapter-midnight@0.0.4
+  - @openzeppelin/contracts-ui-builder-adapter-solana@0.0.3
+  - @openzeppelin/contracts-ui-builder-adapter-stellar@0.0.3
+  - @openzeppelin/contracts-ui-builder-react-core@0.1.3
+  - @openzeppelin/contracts-ui-builder-renderer@0.1.4
+
+## 0.1.3
+
+### Patch Changes
+
+- [#54](https://github.com/OpenZeppelin/contracts-ui-builder/pull/54) [`63fca98`](https://github.com/OpenZeppelin/contracts-ui-builder/commit/63fca981f56bf9b2bb7c43c720bea3cbbd53d6f6) Thanks [@pasevin](https://github.com/pasevin)! - pre-release clean up and improvements
+
+- Updated dependencies [[`63fca98`](https://github.com/OpenZeppelin/contracts-ui-builder/commit/63fca981f56bf9b2bb7c43c720bea3cbbd53d6f6), [`6d74481`](https://github.com/OpenZeppelin/contracts-ui-builder/commit/6d7448140936f5c8dfadac3bca05dde54d468167)]:
+  - @openzeppelin/contracts-ui-builder-adapter-evm@0.1.3
+  - @openzeppelin/contracts-ui-builder-renderer@0.1.3
+  - @openzeppelin/contracts-ui-builder-styles@0.1.2
+  - @openzeppelin/contracts-ui-builder-ui@0.1.3
+  - @openzeppelin/contracts-ui-builder-adapter-midnight@0.0.3
+
 ## 0.1.2
 
 ### Patch Changes