+++ packages/styles/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @openzeppelin/transaction-form-styles
 
+## 0.2.0
+
+### Minor Changes
+
+- [#56](https://github.com/OpenZeppelin/contracts-ui-builder/pull/56) [`83c430e`](https://github.com/OpenZeppelin/contracts-ui-builder/commit/83c430e86f47733bde89b560b70a7a922eebfe81) Thanks [@pasevin](https://github.com/pasevin)! - Support for the new Etherscan V2 Client by the EVM Adapter
+
+## 0.1.2
+
+### Patch Changes
+
+- [#54](https://github.com/OpenZeppelin/contracts-ui-builder/pull/54) [`63fca98`](https://github.com/OpenZeppelin/contracts-ui-builder/commit/63fca981f56bf9b2bb7c43c720bea3cbbd53d6f6) Thanks [@pasevin](https://github.com/pasevin)! - pre-release clean up and improvements
+
 ## 1.16.0
 
 ### Minor Changes